1056 HH spray filled forms changes

+++ opensrp-reveal/src/main/java/org/smartregister/reveal/sync/SprayEventProcessor.java
@@ -2,12 +2,15 @@ package org.smartregister.reveal.sync;
 
 import androidx.annotation.VisibleForTesting;
 
+import net.sqlcipher.Cursor;
 import net.sqlcipher.database.SQLiteDatabase;
 
+import org.apache.commons.lang3.StringUtils;
 import org.smartregister.domain.Client;
 import org.smartregister.domain.Event;
 import org.smartregister.domain.Obs;
 import org.smartregister.domain.jsonmapping.ClientClassification;
+import org.smartregister.repository.EventClientRepository.event_column;
 import org.smartregister.reveal.application.RevealApplication;
 import org.smartregister.reveal.util.Constants.DatabaseKeys;
 import org.smartregister.reveal.util.Constants.Tables;
@@ -17,12 +20,15 @@ import java.util.HashMap;
 
 import static org.smartregister.commonregistry.CommonRepository.ID_COLUMN;
 import static org.smartregister.family.util.DBConstants.KEY.OBJECT_ID;
+import static org.smartregister.reveal.util.Constants.JsonForm.BUSINESS_STATUS;
 import static org.smartregister.reveal.util.Constants.JsonForm.MOP_UP;
 
 /**
  * Created by samuelgithengi on 9/17/20.
  */
 public class SprayEventProcessor {
+
+    private static final String YES = "yes";
 
     private SQLiteDatabase sqLiteDatabase;
 
@@ -52,12 +58,27 @@ public class SprayEventProcessor {
         event.getDetails().put(DatabaseKeys.FORM_SUBMISSION_ID, formSubmissionId);
 
         Obs mopUp = event.findObs(null, true, MOP_UP);
-        if (mopUp != null) {
-            event.setFormSubmissionId(event.getBaseEntityId() + ":" + mopUp.getValue());
+
+        if (mopUp != null && YES.equalsIgnoreCase(mopUp.getValue().toString())) {
+            try (Cursor cursor = getSqLiteDatabase().query(Tables.EVENT_TABLE, new String[]{event_column.json.name()}
+                    , String.format("%s like ? AND %s=?", event_column.baseEntityId, event_column.eventType.name()),
+                    new String[]{event.getBaseEntityId() + "%", event.getEventType()}, null, null, event_column.updatedAt.name() + " DESC", "1")) {
+                if (cursor.moveToNext()) {
+                    Event previousEvent = RevealApplication.getInstance().getContext().getEventClientRepository().convert(cursor.getString(0), Event.class);
+                    Obs previousBusinessStatus = previousEvent.findObs(null, true, BUSINESS_STATUS);
+                    Obs businessStatus = event.findObs(null, true, BUSINESS_STATUS);
+                    if (businessStatus != null && businessStatus.getValue().toString().equalsIgnoreCase(previousBusinessStatus.getValue().toString())) {
+                        event.setFormSubmissionId(event.getBaseEntityId() + ":" + businessStatus.getValue().toString());
+                    } else {
+                        event.setFormSubmissionId(event.getBaseEntityId());
+                    }
+                }
+            }
+
         } else {
             event.setFormSubmissionId(event.getBaseEntityId());
         }
-        if (!isLocalEvent) {
+        if (!isLocalEvent && StringUtils.isNotBlank(event.getBaseEntityId())) {
             getSqLiteDatabase().delete(Tables.EC_EVENTS_TABLE,
                     String.format("%s like ? AND %s=?", ID_COLUMN, DatabaseKeys.EVENT_TYPE),
                     new String[]{event.getBaseEntityId() + "%", event.getEventType()});
@@ -78,4 +99,5 @@ public class SprayEventProcessor {
         }
         return sqLiteDatabase;
     }
+
 }
